requirements-lock.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements-lock.txt requirements.in
#
aistudio-sdk==0.3.8
    # via paddlex
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
    #   openai
bce-python-sdk==0.9.59
    # via aistudio-sdk
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
chardet==5.2.0
    # via paddlex
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via aistudio-sdk
colorama==0.4.6
    # via
    #   -r requirements.in
    #   click
    #   colorlog
    #   loguru
    #   tqdm
coloredlogs==15.0.1
    # via onnxruntime
colorlog==6.10.1
    # via paddlex
distro==1.9.0
    # via
    #   google-genai
    #   openai
filelock==3.20.3
    # via
    #   huggingface-hub
    #   modelscope
    #   paddlex
    #   torch
    #   transformers
fire==0.7.1
    # via
    #   manga-ocr
    #   simple-lama-inpainting
flatbuffers==25.12.19
    # via onnxruntime
fsspec==2026.1.0
    # via
    #   huggingface-hub
    #   torch
fugashi==1.5.2
    # via manga-ocr
future==1.0.0
    # via bce-python-sdk
google-auth[requests]==2.47.0
    # via google-genai
google-genai==1.59.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   openai
    #   paddlepaddle
huggingface-hub==0.36.0
    # via
    #   -r requirements.in
    #   paddlex
    #   tokenizers
    #   transformers
humanfriendly==10.0
    # via coloredlogs
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
imagesize==1.4.1
    # via paddlex
jaconv==0.4.1
    # via manga-ocr
jinja2==3.1.6
    # via torch
jiter==0.12.0
    # via openai
loguru==0.7.3
    # via
    #   -r requirements.in
    #   manga-ocr
manga-ocr==0.1.14
    # via -r requirements.in
markupsafe==3.0.3
    # via jinja2
modelscope==1.34.0
    # via paddlex
mpmath==1.3.0
    # via sympy
natsort==8.4.0
    # via -r requirements.in
networkx==3.6.1
    # via
    #   paddlepaddle
    #   torch
numpy==2.3.5
    # via
    #   -r requirements.in
    #   manga-ocr
    #   onnxruntime
    #   opencv-contrib-python
    #   opencv-python
    #   opt-einsum
    #   paddlepaddle
    #   paddlex
    #   pandas
    #   shapely
    #   simple-lama-inpainting
    #   torchvision
    #   transformers
onnxruntime==1.23.2
    # via -r requirements.in
openai==2.15.0
    # via -r requirements.in
opencv-contrib-python==4.10.0.84
    # via paddlex
opencv-python==4.13.0.90
    # via simple-lama-inpainting
opt-einsum==3.3.0
    # via paddlepaddle
packaging==25.0
    # via
    #   huggingface-hub
    #   onnxruntime
    #   paddlex
    #   transformers
paddleocr==3.3.2
    # via -r requirements.in
paddlepaddle==3.2.2
    # via -r requirements.in
paddlex[ocr-core]==3.3.13
    # via
    #   -r requirements.in
    #   paddleocr
pandas==2.3.3
    # via paddlex
pillow==12.1.0
    # via
    #   manga-ocr
    #   paddlepaddle
    #   paddlex
    #   simple-lama-inpainting
    #   torchvision
prettytable==3.17.0
    # via
    #   aistudio-sdk
    #   paddlex
protobuf==6.33.4
    # via
    #   onnxruntime
    #   paddlepaddle
psutil==7.2.1
    # via aistudio-sdk
py-cpuinfo==9.0.0
    # via paddlex
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pyclipper==1.4.0
    # via paddlex
pycryptodome==3.23.0
    # via bce-python-sdk
pydantic==2.12.5
    # via
    #   google-genai
    #   openai
    #   paddlex
pydantic-core==2.41.5
    # via pydantic
pypdfium2==5.3.0
    # via paddlex
pyperclip==1.11.0
    # via manga-ocr
pyreadline3==3.5.4
    # via humanfriendly
python-bidi==0.6.7
    # via paddlex
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   paddleocr
    #   paddlex
    #   transformers
regex==2026.1.15
    # via transformers
requests==2.32.5
    # via
    #   aistudio-sdk
    #   google-auth
    #   google-genai
    #   huggingface-hub
    #   modelscope
    #   paddleocr
    #   paddlex
    #   transformers
rsa==4.9.1
    # via google-auth
ruamel-yaml==0.19.1
    # via paddlex
safetensors==0.7.0
    # via
    #   paddlepaddle
    #   transformers
shapely==2.1.2
    # via paddlex
simple-lama-inpainting @ git+https://github.com/Mayonnaisu/simple-lama-inpainting
    # via -r requirements.in
six==1.17.0
    # via
    #   bce-python-sdk
    #   python-dateutil
sniffio==1.3.1
    # via
    #   google-genai
    #   openai
sympy==1.14.0
    # via
    #   onnxruntime
    #   torch
tenacity==9.1.2
    # via google-genai
termcolor==3.3.0
    # via fire
tokenizers==0.22.2
    # via transformers
torch==2.9.1
    # via
    #   -r requirements.in
    #   manga-ocr
    #   simple-lama-inpainting
    #   torchvision
torchvision==0.24.1
    # via
    #   -r requirements.in
    #   simple-lama-inpainting
tqdm==4.67.1
    # via
    #   aistudio-sdk
    #   huggingface-hub
    #   modelscope
    #   openai
    #   transformers
transformers==4.57.6
    # via manga-ocr
typing-extensions==4.15.0
    # via
    #   anyio
    #   google-genai
    #   huggingface-hub
    #   openai
    #   paddleocr
    #   paddlepaddle
    #   paddlex
    #   pydantic
    #   pydantic-core
    #   torch
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
ujson==5.11.0
    # via paddlex
unidic-lite==1.0.8
    # via manga-ocr
urllib3==2.6.3
    # via
    #   modelscope
    #   requests
wcwidth==0.2.14
    # via prettytable
websockets==15.0.1
    # via google-genai
win32-setctime==1.2.0
    # via loguru

# The following packages are considered to be unsafe in a requirements file:
setuptools==80.9.0
    # via
    #   modelscope
    #   torch
